requirements.txt
psycopg2-binary==2.9.10
requests==2.32.3
fastapi==0.116.1
uvicorn==0.35.0
